fixed: renamed directory preferences to extensionSettings
