+++ src/commands/docker.ts
@@ -12,10 +12,10 @@ import { ResponseError } from "../clients/docker";
 import { isDockerAvailable } from "../docker/configs";
 import { LocalResourceKind } from "../docker/constants";
 import { LocalResourceWorkflow } from "../docker/workflows/base";
+import { LOCAL_DOCKER_SOCKET_PATH } from "../extensionSettings/constants";
 import { Logger } from "../logging";
 import { ConnectionLabel } from "../models/resource";
 import { showErrorNotificationWithButtons } from "../notifications";
-import { LOCAL_DOCKER_SOCKET_PATH } from "../preferences/constants";
 import { localResourcesQuickPick } from "../quickpicks/localResources";
 import { UserEvent } from "../telemetry/events";
 import { sentryCaptureException } from "../telemetry/sentryClient";
