+++ src/commands/documents.ts
@@ -1,17 +1,17 @@
 import { commands, Disposable, Uri, workspace, WorkspaceConfiguration } from "vscode";
 import { registerCommandWithLogging } from ".";
 import { uriMetadataSet } from "../emitters";
-import { Logger } from "../logging";
-import { CCloudFlinkComputePool } from "../models/flinkComputePool";
-import { CCloudKafkaCluster, KafkaCluster } from "../models/kafkaCluster";
-import { showInfoNotificationWithButtons } from "../notifications";
 import {
   FLINK_CONFIG_COMPUTE_POOL,
   FLINK_CONFIG_DATABASE,
   UPDATE_DEFAULT_DATABASE_FROM_LENS,
   UPDATE_DEFAULT_POOL_ID_FROM_LENS,
-} from "../preferences/constants";
-import { updateDefaultFlinkDatabaseId, updateDefaultFlinkPoolId } from "../preferences/updates";
+} from "../extensionSettings/constants";
+import { updateDefaultFlinkDatabaseId, updateDefaultFlinkPoolId } from "../extensionSettings/updates";
+import { Logger } from "../logging";
+import { CCloudFlinkComputePool } from "../models/flinkComputePool";
+import { CCloudKafkaCluster, KafkaCluster } from "../models/kafkaCluster";
+import { showInfoNotificationWithButtons } from "../notifications";
 import { flinkComputePoolQuickPick } from "../quickpicks/flinkComputePools";
 import { flinkDatabaseQuickpick } from "../quickpicks/kafkaClusters";
 import { hasCCloudAuthSession } from "../sidecar/connections/ccloud";
