+++ src/docker/workflows/base.test.ts
@@ -2,8 +2,8 @@ import * as assert from "assert";
 import * as sinon from "sinon";
 import { window, workspace } from "vscode";
 import { ContainerInspectResponse, ContainerSummary, ResponseError } from "../../clients/docker";
+import { LOCAL_KAFKA_IMAGE } from "../../extensionSettings/constants";
 import { Logger } from "../../logging";
-import { LOCAL_KAFKA_IMAGE } from "../../preferences/constants";
 import * as dockerContainers from "../containers";
 import * as dockerImages from "../images";
 import { LocalResourceContainer, LocalResourceWorkflow } from "./base";
