+++ src/extensionSettings/sidecarSync.test.ts
