+++ src/extensionSettings/sidecarSync.ts
