+++ src/flinkSql/flinkStatementManager.ts
@@ -1,8 +1,5 @@
 import { ConfigurationChangeEvent, Disposable, WorkspaceConfiguration, workspace } from "vscode";
 import { ccloudConnected, flinkStatementDeleted, flinkStatementUpdated } from "../emitters";
-import { CCloudResourceLoader } from "../loaders";
-import { Logger } from "../logging";
-import { FlinkStatement, FlinkStatementId } from "../models/flinkStatement";
 import {
   DEFAULT_STATEMENT_POLLING_CONCURRENCY,
   DEFAULT_STATEMENT_POLLING_FREQUENCY_SECONDS,
@@ -10,7 +7,10 @@ import {
   STATEMENT_POLLING_CONCURRENCY,
   STATEMENT_POLLING_FREQUENCY_SECONDS,
   STATEMENT_POLLING_LIMIT,
-} from "../preferences/constants";
+} from "../extensionSettings/constants";
+import { CCloudResourceLoader } from "../loaders";
+import { Logger } from "../logging";
+import { FlinkStatement, FlinkStatementId } from "../models/flinkStatement";
 import { IntervalPoller } from "../utils/timing";
 import { executeInWorkerPool, extract } from "../utils/workerPool";
