+++ src/quickpicks/utils/schemaSubjects.test.ts
@@ -5,9 +5,9 @@ import { getStubbedLocalResourceLoader } from "../../../tests/stubs/resourceLoad
 import { TEST_LOCAL_SUBJECT } from "../../../tests/unit/testResources/schema";
 import { TEST_LOCAL_SCHEMA_REGISTRY } from "../../../tests/unit/testResources/schemaRegistry";
 import { TEST_LOCAL_KAFKA_TOPIC } from "../../../tests/unit/testResources/topic";
+import { USE_TOPIC_NAME_STRATEGY } from "../../extensionSettings/constants";
 import { LocalResourceLoader } from "../../loaders";
 import { Subject } from "../../models/schema";
-import { USE_TOPIC_NAME_STRATEGY } from "../../preferences/constants";
 import { SubjectNameStrategy } from "../../schemas/produceMessageSchema";
 import * as schemaQuickPicks from "../schemas";
 import { getSubjectNameForStrategy, getSubjectNameStrategy } from "./schemaSubjects";
