+++ src/viewProviders/flinkStatements.test.ts
@@ -7,8 +7,6 @@ import { TEST_CCLOUD_FLINK_COMPUTE_POOL } from "../../tests/unit/testResources/f
 import { createFlinkStatement } from "../../tests/unit/testResources/flinkStatement";
 import { getTestExtensionContext } from "../../tests/unit/testUtils";
 import { flinkStatementDeleted, flinkStatementUpdated } from "../emitters";
-import { CCloudResourceLoader } from "../loaders";
-import { FlinkStatement, Phase } from "../models/flinkStatement";
 import {
   DEFAULT_STATEMENT_POLLING_CONCURRENCY,
   DEFAULT_STATEMENT_POLLING_FREQUENCY_SECONDS,
@@ -16,7 +14,9 @@ import {
   STATEMENT_POLLING_CONCURRENCY,
   STATEMENT_POLLING_FREQUENCY_SECONDS,
   STATEMENT_POLLING_LIMIT,
-} from "../preferences/constants";
+} from "../extensionSettings/constants";
+import { CCloudResourceLoader } from "../loaders";
+import { FlinkStatement, Phase } from "../models/flinkStatement";
 import * as telemetryEvents from "../telemetry/events";
 import { FlinkStatementsViewProvider } from "./flinkStatements";
