+++ src/viewProviders/flinkStatements.ts
@@ -5,11 +5,6 @@ import {
   flinkStatementDeleted,
   flinkStatementUpdated,
 } from "../emitters";
-import { FlinkStatementManager } from "../flinkSql/flinkStatementManager";
-import { CCloudResourceLoader, ResourceLoader } from "../loaders";
-import { CCloudEnvironment } from "../models/environment";
-import { CCloudFlinkComputePool } from "../models/flinkComputePool";
-import { FlinkStatement, FlinkStatementId, FlinkStatementTreeItem } from "../models/flinkStatement";
 import {
   DEFAULT_STATEMENT_POLLING_CONCURRENCY,
   DEFAULT_STATEMENT_POLLING_FREQUENCY_SECONDS,
@@ -17,7 +12,12 @@ import {
   STATEMENT_POLLING_CONCURRENCY,
   STATEMENT_POLLING_FREQUENCY_SECONDS,
   STATEMENT_POLLING_LIMIT,
-} from "../preferences/constants";
+} from "../extensionSettings/constants";
+import { FlinkStatementManager } from "../flinkSql/flinkStatementManager";
+import { CCloudResourceLoader, ResourceLoader } from "../loaders";
+import { CCloudEnvironment } from "../models/environment";
+import { CCloudFlinkComputePool } from "../models/flinkComputePool";
+import { FlinkStatement, FlinkStatementId, FlinkStatementTreeItem } from "../models/flinkStatement";
 import { logUsage, UserEvent } from "../telemetry/events";
 import { BaseViewProvider } from "./base";
